Snap functionality added!
Cards now snap to boxes, every Interactable object has a valid list of things it can snap to and a valid list of things that can snap to it. Additionally, snapping to a box makes a card a child of that box. Picking the card up makes the original parent the parent of the card.